Fixed accidental changes from rebasing

+++ clear/db/migrate/20260429194500_add_confirmable_to_users_copy.rb
@@ -1,10 +1,10 @@
 class AddConfirmableToUsersCopy < ActiveRecord::Migration[8.1]
   def up
-    add_column :users, :confirmation_token, :string
-    add_column :users, :confirmed_at, :datetime
-    add_column :users, :confirmation_sent_at, :datetime
-    add_column :users, :unconfirmed_email, :string
-    add_index :users, :confirmation_token, unique: true
+    add_column :users, :confirmation_token, :string unless column_exists?(:users, :confirmation_token)
+    add_column :users, :confirmed_at, :datetime unless column_exists?(:users, :confirmed_at)
+    add_column :users, :confirmation_sent_at, :datetime unless column_exists?(:users, :confirmation_sent_at)
+    add_column :users, :unconfirmed_email, :string unless column_exists?(:users, :unconfirmed_email)
+    add_index :users, :confirmation_token, unique: true unless index_exists?(:users, :confirmation_token, unique: true)
 
     # So existing users aren't locked out of their accounts
     execute <<~SQL
@@ -15,10 +15,10 @@ class AddConfirmableToUsersCopy < ActiveRecord::Migration[8.1]
   end
 
   def down
-    remove_index :users, :confirmation_token
-    remove_column :users, :unconfirmed_email
-    remove_column :users, :confirmation_sent_at
-    remove_column :users, :confirmed_at
-    remove_column :users, :confirmation_token
+    remove_index :users, :confirmation_token if index_exists?(:users, :confirmation_token)
+    remove_column :users, :unconfirmed_email if column_exists?(:users, :unconfirmed_email)
+    remove_column :users, :confirmation_sent_at if column_exists?(:users, :confirmation_sent_at)
+    remove_column :users, :confirmed_at if column_exists?(:users, :confirmed_at)
+    remove_column :users, :confirmation_token if column_exists?(:users, :confirmation_token)
   end
 end
